update : 네이버 도서 API -> MySQL 구축

+++ src/main/java/com/fastcampus/book_bot/repository/BookRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface BookRepository extends JpaRepository<Book, Long> {
 
-    Optional<Book> findByBookIsbn(Integer bookIsbn);
+    Optional<Book> findByBookIsbn(Long bookIsbn);
 }
